Added more checks

+++ code/server/SKU/SKUdao.js
@@ -38,107 +38,119 @@ class SKUDao {
 
     newSKU(sku) {
         return new Promise((resolve, reject) => {
-            const sql = 'INSERT INTO SKU(DESCRIPTION, WEIGHT, VOLUME, NOTES, POSITION, AVAILABLEQUANTITY, PRICE) VALUES (?, ?, ?, ?, ?, ?, ?)';
-            db.run(sql, [sku.getDescription(), sku.getWeight(), sku.getVolume(), sku.getNotes(), sku.getPosition(), sku.getAvailableQuantity(), sku.getPrice()], (err) => {
-                if (err) {
-                    reject(503);
-                    return;
-                }
-                resolve(sku.get);
-            });
+            let loggedAndAuthorized = true;
+            if (loggedAndAuthorized) {
+                const sql = 'INSERT INTO SKU(DESCRIPTION, WEIGHT, VOLUME, NOTES, POSITION, AVAILABLEQUANTITY, PRICE) VALUES (?, ?, ?, ?, ?, ?, ?)';
+                db.run(sql, [sku.getDescription(), sku.getWeight(), sku.getVolume(), sku.getNotes(), sku.getPosition(), sku.getAvailableQuantity(), sku.getPrice()], (err) => {
+                    if (err) {
+                        reject(503);
+                        return;
+                    }
+                    resolve(sku.get);
+                });
+            } else {
+                console.log('Not logged in or authorized');
+                reject(401);
+            }
         });
     }
 
     getSKUs() {
         return new Promise((resolve, reject) => {
-            const sql = 'SELECT * FROM SKU';
-            db.all(sql, [], (err, rows) => {
-                if (err) {
-                    reject(err);
-                    return;
-                }
-                const skus = rows.map((r) => (
-                    {
-                        id: r.ID,
-                        description: r.description,
-                        weight: r.weight,
-                        volume: r.volume,
-                        notes: r.notes,
-                        position: r.position,
-                        availableQuantity: r.availableQuantity,
-                        price: r.price,
-                        testDescriptors: r.testDescriptors
-                    }));
-                resolve(skus);
-            });
+            let loggedAndAuthorized = true;
+            if (loggedAndAuthorized) {
+                const sql = 'SELECT * FROM SKU';
+                db.all(sql, [], (err, rows) => {
+                    if (err) {
+                        reject(err);
+                        return;
+                    }
+                    const skus = rows.map((r) => (
+                        {
+                            id: r.ID,
+                            description: r.description,
+                            weight: r.weight,
+                            volume: r.volume,
+                            notes: r.notes,
+                            position: r.position,
+                            availableQuantity: r.availableQuantity,
+                            price: r.price,
+                            testDescriptors: r.testDescriptors
+                        }));
+                    resolve(skus);
+                });
+            } else {
+                console.log("Not logged in or wrong permission");
+                reject(401);
+            }
         });
     }
 
     getSKUbyID(id) {
         return new Promise((resolve, reject) => {
-            const checkSKUId = 'SELECT COUNT(*) FROM SKU WHERE ID= ?';
-            let exists = 0;
-            db.all(checkSKUId, [id], (err, res) => {
-                if (err) {
-                    reject(err);
-                    return;
-                }
+            let loggedAndAuthorized = true;
+            if (loggedAndAuthorized) {
+                const checkSKUId = 'SELECT COUNT(*) FROM SKU WHERE ID= ?';
+                let exists = 0;
+                db.all(checkSKUId, [id], (err, res) => {
+                    if (err) {
+                        reject(err);
+                        return;
+                    }
 
-                res[0]['COUNT(*)'] > 0 ? exists = 1 : exist;
+                    res[0]['COUNT(*)'] > 0 ? exists = 1 : exist;
 
-                if (exists) {
-                    const sql = 'SELECT * FROM SKU WHERE ID=?';
-                    db.all(sql, [id], (err, rows) => {
-                        if (err) {
-                            reject(err);
-                            return;
-                        }
-                        const sku = rows.map((r) => (
-                            {
-                                id: r.ID,
-                                description: r.description,
-                                weight: r.weight,
-                                volume: r.volume,
-                                notes: r.notes,
-                                position: r.position,
-                                availableQuantity: r.availableQuantity,
-                                price: r.price,
-                                testDescriptors: r.testDescriptors
+                    if (exists) {
+                        const sql = 'SELECT * FROM SKU WHERE ID=?';
+                        db.all(sql, [id], (err, rows) => {
+                            if (err) {
+                                reject(err);
+                                return;
                             }
-                        ));
-                        resolve(sku[0]);
-                    });
-                } else {
-                    console.log('No SKU associated to ID');
-                    reject(404);
-                }
-            });
+                            const sku = rows.map((r) => (
+                                {
+                                    id: r.ID,
+                                    description: r.description,
+                                    weight: r.weight,
+                                    volume: r.volume,
+                                    notes: r.notes,
+                                    position: r.position,
+                                    availableQuantity: r.availableQuantity,
+                                    price: r.price,
+                                    testDescriptors: r.testDescriptors
+                                }
+                            ));
+                            resolve(sku[0]);
+                        });
+                    } else {
+                        console.log('No SKU associated to ID');
+                        reject(404);
+                    }
+                });
+            } else {
+                console.log('Not logged in or wrong permissions');
+                reject(401);
+            }
         });
     }
 
     modifySKU(id) {
         return new Promise((resolve, reject) => {
-            const checkId = "SELECT * FROM USERS WHERE ID= ?";
-            let exists = 0;
-            db.all(checkId, [id], (err, res) => {
-                if (err) {
-                    reject(err);
-                    return;
-                }
+            let loggedAndAuthorized = true;
+            if (loggedAndAuthorized) {
+                const checkId = "SELECT * FROM USERS WHERE ID= ?";
+                let exists = 0;
+                db.all(checkId, [id], (err, res) => {
+                    if (err) {
+                        reject(err);
+                        return;
+                    }
 
-                res[0]['COUNT(*)'] > 0 ? exists = 1 : exists;
+                    res[0]['COUNT(*)'] > 0 ? exists = 1 : exists;
 
-                if (exists) {
-                    const checkPermission = "SELECT * FROM USERS WHERE ID= ? AND TYPE= 'MANAGER'";
-                    let allowed = 0;
-                    db.all(checkPermission, [id], (err, res) => {
-                        if (err) {
-                            reject(err);
-                            return;
-                        }
-                        res[0]['COUNT(*)'] > 0 ? allowed = 1 : allowed;
-
-                        if (allowed) {
+                    if (exists) {
+                        let loggedAndAuthorized = true;
+                        if (loggedAndAuthorized) {
                             const sql = 'UPDATE SKU SET description = ? ,weight= ?,volume= ?,notes= ?,availablequantity= ?,price= ? WHERE id = ?';
                             db.run(sql, [sku.newDescription, sku.newWeight, sku.newVolume, sku.newNotes, sku.newAvailableQuantity, sku.newPrice, id], (err) => {
                                 if (err) {
@@ -151,40 +163,76 @@ class SKUDao {
                             console.log('SKUID not authorized');
                             reject(401);
                         }
-                    }); 
 
-                } else {
-                    console.log('SKU Id does not exist');
-                    reject(404);
-                }
-            });
+                    } else {
+                        console.log('SKU Id does not exist');
+                        reject(404);
+                    }
+                });
+            } else {
+                console.log('Not logged in or wrong permissions');
+                reject(401);
+            }
 
         });
     }
 
-    modifySKUPosition(id, position) {
+    modifySKUPosition(position) {
         return new Promise((resolve, reject) => {
-            const sql = 'UPDATE SKU SET position = ? WHERE id = ?';
-            db.run(sql, [position, id], (err) => {
-                if (err) {
-                    reject(err);
+            let loggedAndAuthorized = true;
+            if (loggedAndAuthorized) {
+                let checkAvailability = 'SELECT COUNT(*) FROM SKU WHERE POSITION= ?'
+                if (checkAvailability != 0) {
+                    console.log('Position already assigned to a sku');
+                    reject(422);
                 } else {
-                    resolve(true);
+                    const checkVolumeAndWeight = "SELECT S.VOLUME, S.WEIGHT, P.MAXVOLUME, P.MAXWEIGHT, P.OCCUPIEDWEIGHT, P.OCCUPIEDVOLUME FROM SKU S, POSITION P WHERE POSITION = ?"
+                    let checked = false;
+                    db.get(checkVolumeAndWeight, [position], (err, row) => {
+                        if (err) {
+                            reject(err);
+                            return;
+                        }
+                        if ((row.volume <= (row.maxVolume - row.occupiedVolume)) && (row.weight <= (row.maxWeight - row.occupiedWeight)))
+                            checked = true;
+                    });
+                    if (checked) {
+                        const sql = 'UPDATE SKU SET position = ? WHERE id = ?';
+                        db.run(sql, [position], (err) => {
+                            if (err) {
+                                reject(err);
+                            } else {
+                                resolve(true);
+                            }
+                        });
+                    } else {
+                        console.log('Position not capable of satisfying volume and weight constraints');
+                        reject(422);
+                    }
                 }
-            });
+            } else {
+                console.log('Not logged in or wrong permission');
+                reject(401);
+            }
         });
     }
 
     deleteSKUbyID(id) {
         return new Promise((resolve, reject) => {
-            const sql = 'DELETE FROM SKU WHERE id = ?';
-            db.run(sql, [id], (err) => {
-                if (err) {
-                    reject(err);
-                } else {
-                    resolve(true);
-                }
-            });
+            let loggedAndAuthorized = true;
+            if (loggedAndAuthorized) {
+                const sql = 'DELETE FROM SKU WHERE id = ?';
+                db.run(sql, [id], (err) => {
+                    if (err) {
+                        reject(err);
+                    } else {
+                        resolve(true);
+                    }
+                });
+            } else {
+                console.log('Not logged in or wrong permission');
+                reject(401);
+            }
         });
     }
